Операции вынесены в отдельный модуль
